separate GamePadInput config and handlers
