+++ packages/gamepad-input/src/GamepadInput.ts
@@ -1,21 +1,19 @@
 import {
-  getNextEnumValue,
   HedronEngine,
-  InputOptionNodesConfig,
   IPlugin,
   handleEachInput,
-  NodeParamEnum,
   EngineStore,
+  NodeParamEnum,
+  NodeParamBoolean,
+  NodeParamNumber,
+  NodeParamString,
+  NodeParamRGB,
+  NodeParamVector3,
 } from '@hedron/engine'
 import { GamepadManager } from './GamepadManager'
-import {
-  GamepadEvent,
-  GamepadInputType,
-  ShotHandler,
-  ValueHander,
-  ButtonMode,
-  AxisMode,
-} from './GamepadTypes'
+import { GamepadEvent, GamepadInputType, AxisMode, ValueHander, ShotHandler } from './GamepadTypes'
+import { globalOptionNodesConfig, optionNodesConfig } from './GamepadConfig'
+import { createGamepadHandlers, applySmoothing, calculate2AxisValue } from './GamepadHandlers'
 
 /**
  * Plugin to handle gamepad input devices.
@@ -26,268 +24,9 @@ export class GamepadInput implements IPlugin {
   public readonly inputType = 'gamepad'
   public readonly description = 'Handles gamepad input devices.'
   public readonly gamepadManager = new GamepadManager()
-  public readonly globalOptionNodesConfig = [
-    {
-      key: 'axisSmoothing',
-      title: 'Axis Smoothing',
-      valueType: 'number',
-      defaultValue: 0,
-      sliderMin: 0,
-      sliderMax: 0.99,
-    },
-    {
-      key: 'buttonSmoothing',
-      title: 'Button Smoothing',
-      valueType: 'number',
-      defaultValue: 0,
-      sliderMin: 0,
-      sliderMax: 0.99,
-    },
-    {
-      key: 'axisDeadZone',
-      title: 'Axis Dead Zone',
-      valueType: 'number',
-      defaultValue: 0,
-      sliderMin: 0,
-      sliderMax: 0.5,
-    },
-    {
-      key: 'axisCap',
-      title: 'Axis Cap',
-      valueType: 'number',
-      defaultValue: 1,
-      sliderMin: 0,
-      sliderMax: 1,
-    },
-  ] as const satisfies InputOptionNodesConfig
-  public readonly optionNodesConfig = [
-    {
-      key: 'isEnabled',
-      title: 'Enabled',
-      valueType: 'boolean',
-      defaultValue: true,
-    },
-    {
-      key: 'controllerIndex',
-      valueType: 'enum',
-      options: Array.from({ length: 4 }, (_, i) => ({ value: i, label: `Controller ${i + 1}` })),
-      defaultValue: 0,
-    },
-    {
-      key: 'inputType',
-      valueType: 'enum',
-      options: [
-        { value: GamepadInputType.Button, label: 'Button' },
-        { value: GamepadInputType.Axis, label: 'Axis' },
-      ],
-      defaultValue: GamepadInputType.Button,
-    },
-    {
-      key: 'index',
-      valueType: 'enum',
-      options: Array.from({ length: 20 }, (_, i) => ({ value: i, label: `${i}` })),
-      defaultValue: 0,
-    },
-    {
-      key: 'secondaryIndex',
-      title: 'Secondary Index',
-      valueType: 'enum',
-      options: Array.from({ length: 20 }, (_, i) => ({ value: i, label: `${i}` })),
-      defaultValue: 1,
-    },
-    {
-      key: 'triggerOn',
-      valueType: 'enum',
-      options: [
-        { value: 'down', label: 'On Press' },
-        { value: 'up', label: 'On Release' },
-      ],
-      defaultValue: 'down',
-    },
-    {
-      key: 'buttonMode',
-      title: 'Button Mode',
-      valueType: 'enum',
-      options: [
-        { value: ButtonMode.Hold, label: 'Hold' },
-        { value: ButtonMode.Toggle, label: 'Toggle' },
-      ],
-      defaultValue: ButtonMode.Hold,
-    },
-    {
-      key: 'axisMode',
-      title: 'Axis Mode',
-      valueType: 'enum',
-      options: [
-        { value: AxisMode.Single, label: 'Single Axis' },
-        { value: AxisMode.Angle, label: '2 Axis as Angle' },
-        { value: AxisMode.Distance, label: '2 Axis as Distance' },
-      ],
-      defaultValue: AxisMode.Single,
-    },
-    {
-      key: 'angleOffset',
-      title: 'Angle Offset',
-      valueType: 'number',
-      defaultValue: 0,
-      sliderMin: 0,
-      sliderMax: 1,
-    },
-  ] as const satisfies InputOptionNodesConfig
+  public readonly globalOptionNodesConfig = globalOptionNodesConfig
+  public readonly optionNodesConfig = optionNodesConfig
 
-  /**
-   * Handles shot inputs from gamepad events.
-   */
-  private handleShot: ShotHandler = ({ input, engine, gamepadEvent, optionNodes }) => {
-    // Only fire shot on the configured trigger event
-    if (gamepadEvent.inputType === GamepadInputType.Button) {
-      const shouldTrigger =
-        (optionNodes.triggerOn === 'down' && gamepadEvent.isPressed) ||
-        (optionNodes.triggerOn === 'up' && !gamepadEvent.isPressed)
-      if (!shouldTrigger) return
-    }
-    engine.fireShot(input.targetNodeId, { _gamepadEvent: gamepadEvent })
-  }
-
-  /**
-   * Handles enum inputs from gamepad events.
-   */
-  private handleEnum: ValueHander<NodeParamEnum> = ({
-    gamepadEvent,
-    input,
-    storeState,
-    targetNode,
-    optionNodes,
-  }) => {
-    // For buttons, cycle through enum values; for axes, map to enum range
-    if (optionNodes.inputType === GamepadInputType.Button) {
-      const shouldTrigger =
-        (optionNodes.triggerOn === 'down' && gamepadEvent.isPressed) ||
-        (optionNodes.triggerOn === 'up' && !gamepadEvent.isPressed)
-      if (!shouldTrigger) return null
-
-      return getNextEnumValue(input.targetNodeId)(storeState)
-    } else {
-      return targetNode.options[Math.floor(gamepadEvent.value * (targetNode.options.length - 1))]
-        .value
-    }
-  }
-
-  /**
-   * Handles boolean inputs from gamepad events.
-   */
-  private handleBoolean: ValueHander = ({ gamepadEvent, targetNodeValue, optionNodes }) => {
-    if (optionNodes.inputType === GamepadInputType.Button) {
-      // For toggle mode, toggle the state on trigger event
-      if (optionNodes.buttonMode === ButtonMode.Toggle) {
-        const shouldTrigger =
-          (optionNodes.triggerOn === 'down' && gamepadEvent.isPressed) ||
-          (optionNodes.triggerOn === 'up' && !gamepadEvent.isPressed)
-        if (!shouldTrigger) return null
-
-        return !targetNodeValue
-      }
-
-      // For hold mode, follow the button state
-      if (optionNodes.triggerOn === 'down') {
-        // On press: true, on release: false
-        return gamepadEvent.isPressed ?? false
-      } else {
-        // On release: true, on press: false (inverted)
-        return !(gamepadEvent.isPressed ?? true)
-      }
-    } else {
-      return gamepadEvent.value > 0.5
-    }
-  }
-
-  /**
-   * Calculates the combined value for 2-axis modes.
-   */
-  private calculate2AxisValue(
-    primaryValue: number,
-    secondaryValue: number,
-    mode: AxisMode,
-    angleOffset: number,
-  ): number {
-    // Convert from [0, 1] to [-1, 1] for calculation
-    const x = primaryValue * 2 - 1
-    const y = secondaryValue * 2 - 1
-
-    if (mode === AxisMode.Angle) {
-      // Calculate angle using atan2, normalize to [0, 1]
-      const angle = Math.atan2(y, x)
-      const normalizedAngle = (angle + Math.PI) / (2 * Math.PI)
-      // Apply angle offset (wrapping around)
-      return (normalizedAngle + angleOffset) % 1
-    } else if (mode === AxisMode.Distance) {
-      // Calculate distance from origin, clamp to [0, 1]
-      const distance = Math.sqrt(x * x + y * y)
-      return Math.min(1, distance)
-    }
-
-    // Default to primary value
-    return primaryValue
-  }
-
-  /**
-   * Handles number inputs from gamepad events.
-   */
-  private handleNumber: ValueHander = ({ gamepadEvent, storeState, input, optionNodes }) => {
-    const sliderMin = (storeState.nodeValues[`${input.targetNodeId}-sliderMin`] as number) ?? 0
-    const sliderMax = (storeState.nodeValues[`${input.targetNodeId}-sliderMax`] as number) ?? 1
-
-    // For button inputs, check if toggle mode is enabled
-    if (
-      optionNodes.inputType === GamepadInputType.Button &&
-      optionNodes.buttonMode === ButtonMode.Toggle
-    ) {
-      // Only toggle on button down
-      const shouldTrigger =
-        (optionNodes.triggerOn === 'down' && gamepadEvent.isPressed) ||
-        (optionNodes.triggerOn === 'up' && !gamepadEvent.isPressed)
-
-      if (!shouldTrigger) return null
-
-      // Toggle the state
-      const currentToggleState = this.toggleStates.get(input.id) ?? false
-      const newToggleState = !currentToggleState
-      this.toggleStates.set(input.id, newToggleState)
-
-      // Return max if toggled on, min if toggled off
-      return newToggleState ? sliderMax : sliderMin
-    }
-
-    // For axis inputs in 2-axis mode, use the combined value
-    let finalValue = gamepadEvent.value
-    if (
-      optionNodes.inputType === GamepadInputType.Axis &&
-      optionNodes.axisMode !== AxisMode.Single
-    ) {
-      const primaryValue = this.primaryAxisValues.get(input.id) ?? 0.5
-      const secondaryValue = this.secondaryAxisValues.get(input.id) ?? 0.5
-      finalValue = this.calculate2AxisValue(
-        primaryValue,
-        secondaryValue,
-        optionNodes.axisMode,
-        optionNodes.angleOffset,
-      )
-    }
-
-    // Default behavior: map value to slider range
-    return finalValue * (sliderMax - sliderMin) + sliderMin
-  }
-
-  /**
-   * Handles unsupported value types from gamepad events, logging a warning.
-   */
-  private handleUnsupported: ValueHander = ({ input, targetNode, gamepadEvent }) => {
-    console.warn(
-      `Gamepad Input: Unsupported value type for node ${input.targetNodeId}. Value: ${gamepadEvent.value}, Type: ${targetNode.valueType}`,
-    )
-
-    return null
-  }
   store: EngineStore
   engine: HedronEngine
   /**
@@ -298,10 +37,6 @@ export class GamepadInput implements IPlugin {
    * Tracks toggle state for button inputs in toggle mode
    */
   private toggleStates = new Map<string, boolean>()
-  /**
-   * Tracks previous smoothed values for each input target node
-   */
-  private previousValues = new Map<string, number>()
   /**
    * Tracks raw primary axis values for each input (before 2-axis calculation)
    */
@@ -318,18 +53,36 @@ export class GamepadInput implements IPlugin {
   /**
    * Handlers for each input type
    */
-  public handlers = {
-    enum: this.handleEnum,
-    boolean: this.handleBoolean,
-    number: this.handleNumber,
-    string: this.handleUnsupported,
-    rgb: this.handleUnsupported,
-    vector3: this.handleUnsupported,
+  public handlers: {
+    handleShot: ShotHandler
+    enum: ValueHander<NodeParamEnum>
+    boolean: ValueHander<NodeParamBoolean>
+    number: ValueHander<NodeParamNumber>
+    string: ValueHander<NodeParamString>
+    rgb: ValueHander<NodeParamRGB>
+    vector3: ValueHander<NodeParamVector3>
   }
 
   constructor(engine: HedronEngine) {
     this.store = engine.getStore()
     this.engine = engine
+
+    // Create handlers with access to necessary state
+    const baseHandlers = createGamepadHandlers({
+      toggleStates: this.toggleStates,
+      primaryAxisValues: this.primaryAxisValues,
+      secondaryAxisValues: this.secondaryAxisValues,
+    })
+
+    this.handlers = {
+      handleShot: baseHandlers.handleShot,
+      enum: baseHandlers.handleEnum,
+      boolean: baseHandlers.handleBoolean,
+      number: baseHandlers.handleNumber,
+      string: baseHandlers.handleUnsupportedString,
+      rgb: baseHandlers.handleUnsupportedRGB,
+      vector3: baseHandlers.handleUnsupportedVector3,
+    }
 
     this.gamepadManager.onGamepadEvent.add(this.handleGamepadEvent.bind(this))
 
@@ -459,7 +212,7 @@ export class GamepadInput implements IPlugin {
           if (targetNode.nodeType === 'param' && targetNode.valueType === 'number') {
             const primaryValue = this.primaryAxisValues.get(input.id) ?? 0.5
             const secondaryValue = this.secondaryAxisValues.get(input.id) ?? 0.5
-            const combinedValue = this.calculate2AxisValue(
+            const combinedValue = calculate2AxisValue(
               primaryValue,
               secondaryValue,
               optionNodes.axisMode,
@@ -487,7 +240,7 @@ export class GamepadInput implements IPlugin {
           if (isPrimaryAxis) {
             // Use processed value for shots as well
             const processedEvent = { ...event, value: processedValue }
-            this.handleShot({
+            this.handlers.handleShot({
               input,
               engine: this.engine,
               gamepadEvent: processedEvent,
@@ -508,7 +261,7 @@ export class GamepadInput implements IPlugin {
           input,
           storeState,
           optionNodes,
-          // @ts-expect-error -- TS isn't smart enough to infer the correct node type
+          // @ts-expect-error - TS can't infer that handler matches node type
           targetNode,
           targetNodeValue,
         })
@@ -537,46 +290,6 @@ export class GamepadInput implements IPlugin {
 
     // Schedule next update
     window.requestAnimationFrame(() => this.update())
-  }
-
-  /**
-   * Applies smoothing with special handling for angle wrapping.
-   * @param currentValue The current smoothed value
-   * @param targetValue The target value to smooth towards
-   * @param smoothing The smoothing factor (0 = no smoothing, 1 = max smoothing)
-   * @param isAngle Whether this is an angle value that needs wrap-around handling
-   * @returns The smoothed value
-   */
-  private applySmoothing(
-    currentValue: number,
-    targetValue: number,
-    smoothing: number,
-    isAngle: boolean = false,
-  ): number {
-    if (!isAngle) {
-      // Standard linear interpolation
-      return currentValue * smoothing + targetValue * (1 - smoothing)
-    }
-
-    // For angles, handle wrap-around
-    let delta = targetValue - currentValue
-
-    // Normalize delta to [-0.5, 0.5] to take the shortest path
-    if (delta > 0.5) {
-      delta -= 1
-    } else if (delta < -0.5) {
-      delta += 1
-    }
-
-    // Apply smoothing to the delta
-    const smoothedDelta = delta * (1 - smoothing)
-    let result = currentValue + smoothedDelta
-
-    // Wrap result to [0, 1]
-    if (result < 0) result += 1
-    if (result >= 1) result -= 1
-
-    return result
   }
 
   /**
@@ -616,7 +329,7 @@ export class GamepadInput implements IPlugin {
           // Check if this is an angle mode input
           const axisMode = this.inputAxisModes.get(input.id)
           const isAngle = axisMode === AxisMode.Angle
-          finalValue = this.applySmoothing(currentValue, targetValue, smoothing, isAngle)
+          finalValue = applySmoothing(currentValue, targetValue, smoothing, isAngle)
         }
 
         // Update the node value
